xplriot1 board updated to new ubxlib features
The xplriot1 Zephyr board definitions now support the new ubxlib features
which enables performing the necessary board initiations transparently
for the user. This function also handles the switching of uart #2 for
Sara and Nina.
I2C is now enabled by default but can be turned off via cmake option in
case simultaneous use of Sara and Nina is required.